feat: Copy Liked Songs to a playlist with new script
This concludes the build out of a new script to copy liked
songs to a playlist via a script command.

The command replicates the `copy_to_playlist` script but uses
the Liked Songs playlist as the source for the purpose of
creating an archive.

It also adds the documentation for the new command into the
README.